# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Campaign Organizer API
  version: 0.1.0
  description: >
    Personal worldbuilding and RPG campaign management application.
    This document is the **canonical, design-first contract** for the API.
    Both the Spring Boot backend and the React/TypeScript frontend are expected
    to conform to it (see docs/adr/0008-contract-first-openapi.md).

    Phase 0 covers single-password authentication and the `worlds` resource,
    which is the top-level container every other domain object will hang off of.
  license:
    name: Proprietary (personal use)
  contact:
    name: Campaign Organizer
servers:
  - url: /api
    description: Same-origin API root (frontend proxies /api to the backend)
...
paths:
...
  /worlds:
    get:
      tags: [worlds]
      summary: List all worlds
      operationId: listWorlds
      responses:
        '200':
          description: All worlds owned by the user, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/World'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [worlds]
      summary: Create a world
      operationId: createWorld
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldRequest'
      responses:
        '201':
          description: World created.
          headers:
            Location:
              description: URI of the created world.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /worlds/{worldId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
    get:
      tags: [worlds]
      summary: Get a single world
      operationId: getWorld
      responses:
        '200':
          description: The world.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [worlds]
      summary: Update a world
      operationId: updateWorld
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldRequest'
      responses:
        '200':
          description: World updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [worlds]
      summary: Delete a world
      operationId: deleteWorld
      responses:
        '204':
          description: World deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worlds/{worldId}/maps/{mapId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/MapId'
    get:
      tags: [maps]
      summary: Get a map
      operationId: getMap
      responses:
        '200':
          description: The map.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Map'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [maps]
      summary: Update a map
      operationId: updateMap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapRequest'
      responses:
        '200':
          description: Map updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Map'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [maps]
      summary: Delete a map (and its pins)
      operationId: deleteMap
      responses:
        '204':
          description: Map deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /worlds/{worldId}/map-categories:
    parameters:
      - $ref: '#/components/parameters/WorldId'
    get:
      tags: [maps]
      summary: List a world's map categories (ADR-0105, separate from Wiki's categories)
      operationId: listMapCategories
      responses:
        '200':
          description: Map categories in the world, ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MapCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [maps]
      summary: Create a map category
      operationId: createMapCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapCategoryRequest'
      responses:
        '201':
          description: Map category created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /worlds/{worldId}/map-categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/CategoryId'
    put:
      tags: [maps]
      summary: Update a map category
      operationId: updateMapCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapCategoryRequest'
      responses:
        '200':
          description: Map category updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [maps]
      summary: Delete a map category (its maps are kept, uncategorized)
      operationId: deleteMapCategory
      responses:
        '204':
          description: Map category deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token obtained from POST /auth/login. Signed with HS256.
  parameters:
    WorldId:
      name: worldId
      in: path
      required: true
      description: Identifier of the world.
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: Identifier of the category.
      schema:
        type: string
        format: uuid
...
    MapId:
      name: mapId
      in: path
      required: true
      description: Identifier of the map.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid bearer token / password.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    WorldRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          examples: ["Aethelgard"]
        description:
          type: string
          maxLength: 5000
          description: Optional summary of the world.
        scratch:
          type: boolean
          default: false
          description: >-
            Cosmetic-only sandbox/brainstorming flag (FR-60) — no functional
            effect, just a label in the world list and switcher.
    World:
      type: object
      required: [id, name, scratch, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        layerStyles:
          $ref: '#/components/schemas/LayerStyleMap'
        scratch:
          type: boolean
          description: >-
            Cosmetic-only sandbox/brainstorming flag (FR-60) — no functional
            effect, just a label in the world list and switcher.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LayerStyle:
      type: object
      description: Styling for a map layer.
      properties:
        color:
          type: [string, "null"]
        icon:
          type: [string, "null"]
    LayerStyleMap:
      type: object
      description: Per-layer styling keyed by layer name.
      additionalProperties:
        $ref: '#/components/schemas/LayerStyle'
...
    MapRequest:
      type: object
      required: [name, mediaId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        mediaId:
          type: string
          format: uuid
          description: The uploaded image used as the map's base layer.
        categoryId:
          type: [string, "null"]
          format: uuid
          description: Map category for the Atlas tree (ADR-0105), or null for uncategorized.
    Map:
      type: object
      required: [id, worldId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        worldId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        mediaId:
          type: [string, "null"]
          format: uuid
        imageUrl:
          type: [string, "null"]
          description: Public URL of the base image (null if the image was deleted).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MapCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parentId:
          type: [string, "null"]
          format: uuid
          description: Parent map category for nesting, or null for a root category.
    MapCategory:
      type: object
      required: [id, worldId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        worldId:
          type: string
          format: uuid
        parentId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      description: RFC 9457 / RFC 7807 problem response.
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
security:
  - bearerAuth: []